backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel>=0.0.22
pydantic>=2
pymysql
python-dotenv